Add caching infrastructure: introduce configuration, repository interfaces, database context, and implementation for translation and chat history management.

+++ TataruLink/Pipeline/Stages/Display/DisplayStage.cs
@@ -79,7 +79,7 @@ public class DisplayStage(TataruConfig configuration, IDataService dataService, 
             {
                 MessageId = message.Id,
                 Timestamp = DateTimeOffset.UtcNow.ToUnixTimeSeconds(),
-                ChatType = (ushort)message.ChatType,
+                ChatType = message.ChatType,
                 ChatTypeName = message.GetChannelName(),
                 SenderName = message.SenderName,
                 OriginalContent = message.PlainTextContent,
